Added functionality to delete a book.
Allows a user to delete a book from the database by using the title of the book. Additional checks will be required to be added to ensure the book is in the database. Additionally, some natural language processing could occur to try to use auto-correct to fix any human error.